Fix linter errors

+++ x/slashing/keeper_test.go
@@ -165,7 +165,7 @@ func TestHandleAbsentValidator(t *testing.T) {
 	validator, _ := sk.GetValidatorByPubKey(ctx, val)
 	require.Equal(t, sdk.Bonded, validator.GetStatus())
 	pool := sk.GetPool(ctx)
-	require.Equal(t, int64(amtInt), pool.BondedTokens.RoundInt64())
+	require.Equal(t, amtInt, pool.BondedTokens.RoundInt64())
 
 	// 501st block missed
 	ctx = ctx.WithBlockHeight(height)
@@ -195,7 +195,7 @@ func TestHandleAbsentValidator(t *testing.T) {
 	// validator should have been slashed
 	pool = sk.GetPool(ctx)
 	slashAmt := sdk.NewDec(amtInt).Mul(keeper.SlashFractionDowntime(ctx)).RoundInt64()
-	require.Equal(t, int64(amtInt)-slashAmt, pool.BondedTokens.RoundInt64())
+	require.Equal(t, amtInt-slashAmt, pool.BondedTokens.RoundInt64())
 
 	// validator start height should have been changed
 	info, found = keeper.getValidatorSigningInfo(ctx, sdk.ConsAddress(val.Address()))
@@ -252,7 +252,7 @@ func TestHandleNewValidator(t *testing.T) {
 
 	info, found := keeper.getValidatorSigningInfo(ctx, sdk.ConsAddress(val.Address()))
 	require.True(t, found)
-	require.Equal(t, int64(keeper.SignedBlocksWindow(ctx)+1), info.StartHeight)
+	require.Equal(t, keeper.SignedBlocksWindow(ctx)+1, info.StartHeight)
 	require.Equal(t, int64(2), info.IndexOffset)
 	require.Equal(t, int64(1), info.SignedBlocksCounter)
 	require.Equal(t, time.Unix(0, 0).UTC(), info.JailedUntil)
@@ -296,7 +296,7 @@ func TestHandleAlreadyJailed(t *testing.T) {
 	require.Equal(t, sdk.Unbonding, validator.GetStatus())
 
 	// validator should have been slashed
-	require.Equal(t, int64(amtInt-1), validator.GetTokens().RoundInt64())
+	require.Equal(t, amtInt-1, validator.GetTokens().RoundInt64())
 
 	// another block missed
 	ctx = ctx.WithBlockHeight(height)
@@ -304,6 +304,6 @@ func TestHandleAlreadyJailed(t *testing.T) {
 
 	// validator should not have been slashed twice
 	validator, _ = sk.GetValidatorByPubKey(ctx, val)
-	require.Equal(t, int64(amtInt-1), validator.GetTokens().RoundInt64())
+	require.Equal(t, amtInt-1, validator.GetTokens().RoundInt64())
 
 }
